Gathers up lint ignore functions under an Ignorer object
